Update to Simulator, MQTT unused port. See description
update to make configurable the number of animals that are created on sim start up (currently there is only ever one animal simulated on each start up).

Also, the Port 7000 for MQTT can not be used on MacOS as there is already an OS service binding to that port, so changes MQTT to bind to 7001 for some service.

Also removed docker-compose file for MQTT as it is not required.