EE/DLe/Librosa-Image-Display
Added Jupyter notebook including a function to display a MEl-spec image, and a wave function image.
User can use the function to dictate if the x and y axis labels, title, bar graph key, and image shape should be displayed. Functions also allow the title and image width/height to be entered as custom values