EE/AKk/Adding/CAM/to/Spectograms
I have made the proposed changes and added everything to the readme file and modified the ipynb notebook accordingly.

The image_folder is required to store the spectograms created by the plot mel spec only notebook shared by Michael, I have added the notebook in the folder so it is easier to access for further use.

The atleast 3 images part is to plot the spectograms with overlays and CAM's in a 3x3 grid for 3 different species as requested by Michael.