EE/DP/Random_Sampling
Update README>md
This code will allow you to randomly select the number of files needed for testing. The Input directory has to be provided from where the folders will be selected and the files within them will be randomly selected and stored in the output folder with the same folder name.