Species select (Show or Hide on the ol map) (#453)
* Added search bar in filter menu

Added search bar in filter menu that updates real time to the text entered in the search bar. Gathered the animal names from the sample_data.json file

* Added summary popup

Adedd popup window for when a list item is clicked in the search

* Show and Hide buttons work in filter

Show and hide buttons in the filter functionality work per animal. However there is a bug of the animals not coming back with the right icon color

* Fixed incorrect colour icons

Fixed incorrect colour icons showing up once hidden and displayed again

* Added show only and unhide all functionality

Added two buttons that allow users to only show a certain species and to go back to showing all species

* Added a filter list to show specific animals

Updated logic to show multiple specific animals and display those in a list.

* Added hidden animals list

Added functionality to hide multiple animals together and they show up in the hidden animals list. The hidden animals list and the shown only list work together now.

* Added protection against Cross-site scripti

Added sanitization to search bar and moved css from the html to the relevant css file

---------

Co-authored-by: gokhanpicgeta <vishnuprasad97@hotmail.com>