model metrics
i used python to compare all the outputs in the respective notebooks for each model based on the training/epochs. I need to now work on testing on some more test data to get further metrics but this is solid for now. In talks with tanmay via teams to get this sorted.