For submission overview
The submissions file stores details about audio submissions made by users on the Echo Map. Each entry in the file includes a unique identifier for the submission, the name or URL of the audio file, the submission date, a list of animals detected in the audio, the user who submitted the file, and additional metadata such as the duration and quality of the audio. This file is essential for displaying the submission information on the Submissions Overview Page and is regularly updated with new data as users upload audio files. It serves as the central data structure for managing and retrieving submission details throughout the project.

Signed-off-by: mariaelsa14 <167952119+mariaelsa14@users.noreply.github.com>