Completed Sprint 2 Work on Benchmarking_Framework
Significant updates to V2 of Benchmarking framework notebook. This can now be used to train models and run experiments detailed in the relevant config files. Models will be exported to the 'models' directory.

Future work includes creating code for visualisation of model training for comparison, and confirmation of repeatability of experiments.