Add experiment runner and visualization utilities
This is a comprehensive update to the B&E framework, which also now includes comprehensive instructions on how to run this notebook. It has been cleaned to make it easier to read and use. Many of the functions have been moved to separate python files for re-usability and enhance readability